requirements.txt
requests>=2.25.1
streamlit>=1.0.0
pandas>=1.1.0
matplotlib>=3.3.0
seaborn>=0.11.0
firecrawl>=0.0.1  # Adjust the version according to the actual one used
python-dotenv>=0.19.0
beautifulsoup4>=4.9.3
textstat>=0.7.1
